backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.17
email-validator==2.2.0
aiofiles==24.1.0
filetype==1.2.0
slowapi==0.1.9
python-json-logger==3.2.1
prometheus-fastapi-instrumentator==7.0.0
